pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sm-mlops-autoretrain-template"
version = "0.1.0"
description = "End-to-end ML + MLOps template with AWS SageMaker auto-retraining."
authors = [{ name = "Template Author" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.34.0",
    "sagemaker>=2.214.1",
    "typer>=0.12.3",
    "pyyaml>=6.0.1",
    "pandas>=2.2.2",
    "numpy>=1.26.4",
    "scikit-learn>=1.4.2",
    "torch>=2.2.2",
    "matplotlib>=3.8.4",
    "pyarrow>=15.0.2",
    "fastparquet>=2024.2.0",
    "rich>=13.7.1",
    "retrying>=1.3.4",
    "jinja2>=3.1.3",
    "pytest>=8.1.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.7",
    "black>=24.3.0",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "S", "TID", "ERA"]
ignore = ["E203", "E266"]
exclude = ["build", "dist"]

[tool.black]
line-length = 100
target-version = ["py311"]
